Introduce VirtualEnv class
This commit introduces a class for VirtualEnv which keeps an internal
reference to the virtualenv destination dir. Also note that this
directory is no longer configurable, but instead is safely tucked away
in a temporary directory.

The virtualenv singleton has also been changed to remove all functions
other than create(), which now returns a new instance of the
VirtualEnv class.