Introduce VirtualEnv.create() method
Because we can't safely create the virtualenv in the constructor, this
work was previously done in the the corresponding singleton for this
class. However, we can't easily test such singletons, and furthermore,
they should really do as little work as possible. Thus, we now move
the virtualenv creation call to its own method within the VirtualEnv
class.