Introduce new class: PythonBuilder
The PythonBuilder class will download a Python tarball from the
internet, extract it to the job's temporary directory, compile it, and
then return a string to the installation directory (also located in
the job's temporary directory). This takes about two minutes on a
modestly-powered Linux node, but can be sped up with the 'makeJobs'
argument.

A corresponding singleton has also been for easy access from
Jenkinsfiles.