Allow Python versions to be loaded via pyenv
Overloads the create() method of VirtualEnv to allow passing a new
parameter pointing to the root of the pyenv installation. If present,
a virtualenv will be created using the Python version specified by the
"python" parameter. This string should match a valid Python version as
it would be passed to "pyenv install".

Co-authored-by: nre <nre@ableton.com>