Store virtualenvs in workspace, not temp folder
This feature was added in 5eaa6088 because the workspace directories
created by the GitHub Organization plugin could be very long, and
would cause problems since the length of the script shebang is limited
to 128 characters.

This change may cause problems for these users; hence, it is an API
breaking change and a version bump will be introduced in a subsequent
commit.