Make Pyenv.createVirtualEnv non-static
This interface makes more sense, because we are creating a Pyenv object
with the given parameters. The virtualenv doesn't know (or care) about
the pyenv root. However, the singleton interface is the same, but
internally, this is a bit nicer.

In subsequent commits, we'll add some more checks that were awkward or
difficult to do when this factory method lived under the VirtualEnv
class.