Fail with error if version not supported by Pyenv
Previously, if the user tried to install a Python version that Pyenv
didn't support, this would result in a very long error message, mostly
because the `sh` step in `createVirtualEnv` contains a number of
commands, so the output for all of them is present in this step.

Granted, the error clearly explains what happens, but it's a lot of text
to wade through to get to the root cause.