Fix test for Pyenv.createVirtualEnv
Previously, we only cared that this method didn't throw, but that led to
the situation where a misconfigured mock meant that this test silently
didn't do anything. This commit fixes that mock and makes the test
actually ensure that the VirtualEnv was created correctly.

Should have been done in d4fe862.

Make test for createVirtualEnv assert something