Run pyenv to assert the pyenv root path exists
To support Windows (which will be done in a subsequent commit), we'll
need to translate Windows-style paths to Cygwin-style paths (for
example, C:\foo\bar -> /c/foo/bar). The fileExists step only supports
Windows-style paths when run on Windows, which means that we need to
actually run pyenv to assert that it exists.