Cleanup: make installCommands method take a map
Two parameters with default values is madness, granted this is a private
helper function in a test class, but it makes for some test code that is
hard to read. In a subsequent commit we'll add other arguments to this
method and that would be the straw that breaks the camel's back.