Don't prepend to existing PATH in Pyenv operations
This addresses an issue in macOS where Python versions from Homebrew or
MacPorts can interfere with Pyenv installation operations. But moreover,
it ensures that we are only going to be able to access Pyenv and the
base system tools, which will hopefully prevent similar errors like this
in the future.

On Windows, we still need the full `$PATH` to be able to find `cmd.exe`,
so this change is only active for Unix nodes.